# DOC/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto controle de filmes - YASLA
  description: |-
   API destinada ao controle de filmes para o projeto bem legal do SENAI
  termsOfService: http://swagger.io/terms/
  contact:
    email: franschesco64@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io 
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: Endpoint para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo Filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internal Server (Controller/Model)    
    get:
      tags:
        - Filme
      summary: Lista todos os Filmes
      description: Listagem de todos os Filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Controller/Model)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um Filme pelo ID
      description: Retorna um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Filme not found
        '500':
          description: Error Internal Server (Controller/Model)
    put:
      tags:
        - Filme
      summary: Atualiza um filme pelo id
      description: Buscar e atualizar um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um Filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found     
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internal Server (Controller/Model)  
    delete:
      tags:
        - Filme
      summary: Excluir um Filme pelo ID
      description: Buscar e excluir um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '500':
          description: Error Internal Server (Controller/Model)
components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
        - data_lancamento
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
          format: date
          example: "2026-05-15"
        duracao:
          type: string
        capa:
          type: string
        valor:
          type: number
          format: double
          example: "999.99"
          description: Valor em formato decimal de até 10 digitos
        avaliacao:
          type: integer
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    